src/codeguard-mcp/pyproject.toml
[project]
name = "codeguard-mcp"
version = "0.1.0"
description = "MCP Server exposing CoSAI CodeGuard security rules as tools"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
keywords = ["mcp", "security", "codeguard", "cosai", "owasp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
]

[project.scripts]
codeguard-mcp = "codeguard_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codeguard_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
